+++ tests/Fixtures/TestController.php
@@ -1,0 +1,11 @@
+<?php
+
+class TestController
+{
+
+    public function validate(TestForm $form)
+    {
+
+    }
+    
+}
