+++ tests/FormBuilderValidationTest.php
@@ -7,29 +7,6 @@ namespace {
     use Kris\LaravelFormBuilder\Form;
     use Kris\LaravelFormBuilder\FormBuilder;
     use Kris\LaravelFormBuilder\FormHelper;
-    use Kris\LaravelFormBuilder\Traits\ValidatesWhenResolved;
-
-    class TestForm extends Form
-    {
-        use ValidatesWhenResolved;
-
-        public function buildForm()
-        {
-            $this->add('name', 'text', ['rules' => ['required', 'min:3']]);
-            $this->add('email', 'text', ['rules' => ['required', 'email', 'min:3']]);
-        }
-
-    }
-
-    class TestController
-    {
-
-        public function validate(TestForm $form)
-        {
-
-        }
-
-    }
 
     class FormBuilderValidationTest extends FormBuilderTestCase
     {
